==> 56_models_final.ipynb <==
# 56. Models (final)
## Contents
- Prerequisites
- Results
    - Logistic Regression
    - Random Forest
    - Support Vector Machine
- Final Model
---------------------------------------------------------
## Prerequisites

import time
import os
import pandas as pd
import numpy as np
from sklearn.dummy import DummyClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from tqdm import tqdm
import joblib
import warnings
from sklearn.model_selection import cross_val_score
warnings.filterwarnings("ignore", category=FutureWarning)
warnings.filterwarnings("ignore", category=UserWarning)
file_path = 'C:/Users/Koen.Janssen/OneDrive/Documents/TILBURG UNIVERSITY/Thesis/Data/'
parties = np.load('00_parties.npy')
options = ['imbalanced','oversampling','undersampling','balancedsampling']
rs = 1

------------------------------------------------
<br>
<br>
<br>


## Results
### Logistic Regression
--------------------------------------------

file_url  = file_path + '53_data_tuned_logistic_regression.csv' 
result_lr = pd.read_csv(file_url)
columns = ['option', 'C', 'max_iter']
result_lr = result_lr.drop(['option'], axis=1)
result_lr.columns = result_lr.columns.map(lambda x: 'lr_' + x if x in columns else x)

avg_best_lr = result_lr['lr_score'].mean()                                             # define average best values
print("Average accuracy: ", avg_best_lr)                                               # print average best values

display(result_lr)

------------------------------------
### Random Forest

file_url  = file_path + '54_data_tuned_random_forest.csv' 
result_rf = pd.read_csv(file_url)
columns = ['option', 'n_est', 'max_depth','min_split']
result_rf = result_rf.drop(['option'], axis=1)
result_rf.columns = result_rf.columns.map(lambda x: 'rf_' + x if x in columns else x)

avg_best_rf = result_rf['rf_score'].mean()                                             # define average best values
print("Average accuracy: ", avg_best_rf)                                               # print average best values

display(result_rf)

-------------------------
### Support Vector Machine

file_url  = file_path + '55_data_tuned_support_vector_machine.csv' 
result_svm = pd.read_csv(file_url)
columns = ['option', 'C', 'kernel','gamma']
result_svm = result_svm.drop(['option'], axis=1)
result_svm.columns = result_svm.columns.map(lambda x: 'svm_' + x if x in columns else x)

avg_best_svm = result_svm['svm_score'].mean()                                          # define average best values
print("Average accuracy: ", avg_best_svm)                                              # print average best values

display(result_svm)

--------------------------

result_all = pd.merge(result_lr, result_rf, on='party', how='inner')
result_all = pd.merge(result_all, result_svm, on='party', how='inner')
display(result_all)

result_best = result_all.copy()
result_best['best_score'] = result_best[['lr_score', 'rf_score', 'svm_score']].idxmax(axis=1)
result_best = result_best[['party', 'best_score', 'lr_score', 'rf_score', 'svm_score']]
print("Average accuracy: ", avg_best_lr)                                               # print average best values
print("Average accuracy: ", avg_best_rf)                                               # print average best values
print("Average accuracy: ", avg_best_svm)                                              # print average best values
display(result_best)

------------------------------------------------
<br>
<br>
<br>


## Final Models
--------------------------------------------

dur = time.time()
fm_results = pd.DataFrame(columns=['party','n_est', 'max_depth', 'min_split','fm_score'])
file_url  = file_path + '54_data_tuned_random_forest.csv' 
parameters = pd.read_csv(file_url)
parameters = parameters.drop(['option','rf_score'], axis=1)

# -------------------------------------------------------------------------------------~----------------------------------------
for party in tqdm(parties, desc = 'Party loop'):                                       # for each party
    
# load training and validation data 
# -------------------------------------------------------------------------------------~----------------------------------------
    X_trainvalid_url = file_path + f"40_models_imbalanced/" + f"{party}_X_trainvalid.csv"
    y_trainvalid_url = file_path + f"40_models_imbalanced/" + f"{party}_y_trainvalid.csv"
    X_trainvalid = pd.read_csv(X_trainvalid_url)                                       # read X_trainvalid
    X_trainvalid = X_trainvalid.drop(columns=['source', 'text','stemming_id','document_id'])
    y_trainvalid = pd.read_csv(y_trainvalid_url)                                       # read y_trainvalid
        
# load test data
# -------------------------------------------------------------------------------------~----------------------------------------
    X_test_url = file_path + f"40_models_imbalanced/" + f"{party}_X_test.csv"          # define X_test url
    y_test_url = file_path + f"40_models_imbalanced/" + f"{party}_y_test.csv"          # define y_test url
    X_test = pd.read_csv(X_test_url)                                                   # read X_test
    X_test = X_test.drop(columns=['source', 'text','stemming_id','document_id'])       # redefine X_test without columns
    y_test = pd.read_csv(y_test_url)                                                   # read y_test

# Define best model (parameters)
# -------------------------------------------------------------------------------------~----------------------------------------
    party_parameters = parameters[parameters['party'] == party].iloc[0]
    n_est = party_parameters['n_est']
    max_depth = party_parameters['max_depth']
    min_split = party_parameters['min_split']    

# Run best model 
# -------------------------------------------------------------------------------------~----------------------------------------
    # -----------------------------------------
    fm = RandomForestClassifier(n_estimators=n_est, max_depth=max_depth, min_samples_split=min_split, random_state=rs)
    fm.fit(X_trainvalid, y_trainvalid)                                                 # fit model
    fm_score = fm.score(X_test, y_test)                                                # score model
    fm_result = pd.DataFrame({  'party': [party],
                                'n_est': [n_est],
                                'max_depth': [max_depth],
                                'min_split': [min_split],
                                'fm_score': [fm_score]})                               # define result
        
    fm_results = pd.concat([fm_results, fm_result], ignore_index=True)                 # redefine results
    model_url = file_path + f"50_models/{party}.pkl"                                   # define model url
    joblib.dump(fm, model_url)                                                         # dump model

# -------------------------------------------------------------------------------------~----------------------------------------
display(fm_results)                                                                    # display results
avg_best_fm = fm_results['fm_score'].mean()                                            # define average best values
print("Average accuracy: ", avg_best_fm)                                               # print average best values
print('\n---------------------------------------------------------------------------------------------------------------------')
print(f"Code duration: {round((time.time()  - dur),3)} seconds")
